moved dir around, reorganize front end code(react) seaprate fromm backend code(express), and updated some path name in source files
