--- src/myapp/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "opensource-ny",
-  "version": "1.0.0",
-  "description": "a tool that tracks contributions made by a participant on github",
-  "main": "server.js",
-  "scripts": {
-    "test": "jest --detectOpenHandles",
-    "start": "node server.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/yizongk/OpenSource-NY.git"
-  },
-  "author": "yizongk gutierrezjdr mxmsunny Undid-Iridium Chocolate-Spaghet rramnauth11",
-  "license": "SEE LICENSE IN LICENSE.md",
-  "bugs": {
-    "url": "https://github.com/yizongk/OpenSource-NY/issues"
-  },
-  "homepage": "https://github.com/yizongk/OpenSource-NY#readme",
-  "dependencies": {
-    "express": "^4.16.4"
-  },
-  "devDependencies": {
-    "jest": "^24.5.0",
-    "supertest": "^4.0.0"
-  },
-  "jest": {
-    "verbose": true,
-    "collectCoverage": true,
-    "roots": [
-      "test/"
-    ]
-  }
-}
